Split the settings-dialog helpers out of menu.mjs
menu.mjs carried the menu bar, the tab helpers, the toolbar and the settings
dialog, and had a standing transition exception at 345 lines saying to split
the dialog helpers next. Adding one assertion to it took the file to 346 and
the guard called that in.

assertSettingsDialog, openSettingsDialog and closeSettingsDialog moved to
settingsDialog.mjs. menu.mjs is 250 lines and no longer needs an exception at
all, so the entry is gone from the guard rather than raised.

Callers import the dialog helpers from the module that owns them. Re-exporting
them through menu.mjs would have left a cycle -- settingsDialog needs
waitForNoElement from menu -- which ESM tolerates for hoisted declarations
but is worth not having.
